refact: 중복되는 모달 로직을 분리합니다.
